+++ src/FunFair.BuildCheck.ProjectChecks/Settings/PackageTagsMetadata.cs
@@ -1,0 +1,15 @@
+using Microsoft.Extensions.Logging;
+
+namespace FunFair.BuildCheck.ProjectChecks.Settings
+{
+    /// <summary>
+    /// Checks that the &quot;PackageTags&quot; property is set in the project.
+    /// </summary>
+    public sealed class PackageTagsMetadata : PackableMetadataBase
+    {
+        public PackageTagsMetadata(ILogger<PackageTagsMetadata> logger)
+            : base(@"PackageTags", logger)
+        {
+        }
+    }
+}